Implement new logical key fetch pipeline stub
We no longer simply keep track of all the pressed key ids in an array.
Standard and custom keys are now distinguished and kept separate.

Custom keys go trough a substitution phase (currently a stub) which
lets us do things like substituting `LAYER3` when pressing `LAYER1`
and `LAYER2` together.

After the substitution phase, the desired layer is computed by looking
at the substituted keys and then in the custom keys array, in this
order.

After we know which layer we should be looking at, we can then fetch
all the relevant logical keys and produce the correct usb report.

Please note that, for usability purposes, the modifier keys are
inherited by all non-base layers, this means that if a key is a
modifier on the base layer (for example shift) the same key will
be the same modifier on all other non-base layers.

On the other hand, it is possible to define additional modifier
keys on non-base layers if so desired.